Overhaul of the cog system
