+++ .gitignore
@@ -1,13 +1,4 @@
 __pycache__/
 *.pyc
-config/config.py
 config/config.ini
-config/emotes.py
 .vscode/
-mods/testmods/
-data/
-mods/memberEvents.py
-mods/manbot.py
-mods/info.py
-mods/info.py
-*.py
