+++ config/config.example.ini
@@ -11,9 +11,3 @@ Status = with Aleana!
 
 # Sets the status to Playing (0), Listening (2), or Watching (3).
 Activity = 0
-
-# Allows access to the Osu API for Osu commands.
-OsuAPI = 
-
-# This is the API key for Push Bullet integration.
-PushBulletToken =
