--- config/config.example.py
@@ -1,11 +1,0 @@
-# Modules for commands go here.
-# Do *NOT* mess with this if you don't know what you're doing.
-# This is here for those that would like to make additions and customize their bot,
-# but can easily prevent you from using commands if you delete the existing modules.
-Modules = [
-    'mods.cogloader',
-    'mods.manbot',
-    'mods.commands',
-    'mods.moderate',
-    'mods.info'
-]
