--- mods/commands.py
@@ -1,28 +1,0 @@
-import asyncio
-import discord
-import math
-import inspect
-import datetime
-import aiohttp
-import utils
-import json
-import requests
-from discord import Game
-from discord.ext import commands
-from discord.ext.commands import Bot
-from config import config
-from utils.embed import Embeds
-from utils.config import Config
-
-class Commands(commands.Cog):
-    def __init__(self, bot):
-        self.bot = bot
-
-    @commands.command(name='help',
-                hidden=True)
-    async def help(self, ctx):
-        emb = Embeds.create_embed(self, ctx, "Help", color=0x00ff00, message="https://aleanaazure.xyz/doku.php?id=alebot")
-        await ctx.send(embed=emb)
-
-def setup(bot):
-    bot.add_cog(Commands(bot))
